Added functionalities to checklist screen
I made it so when you create an item in the checklist screen, the name and quantity you enter in the modal actually show up as the name and quantity of the item created. I also added increment and decrement quantity for the items on the checklist, and it works the same way as the increment and decrement quantity for the food screens.