Firebase realtime database functionalities
Added the ability to add items to the realtime database from each of the food screens.

Made it so the checklist screen displays the contents of the realtime database and automatically re-renders when any changes are made to the contents. And also on the checklist screen, you can now add items to the database, delete items from the database, and increment/decrement the quantity of items in the database.